Fix used placing of the snap placing
Now rendered meshes of the origin and the attached object
in the process of place snapping math up no matter if meshes
are different

Adjusted GenerateHallways component und Tool for snap spawning
to the new api in SnapMover.

Deleted SnapSpawnerComponent. LevelManger can access SnapMover
directly. And no use for a separate component so far.

Added comments in SnapSpawnController, SpawnSnapWindow, SpawnSnapWindow
, SnapSpawnerComponent, GenerateHallways, ApplyRotationScaleState.